Explicitly set strategy on factory associations
`factory_bot` versions 5.x and newer change the default for
`use_parent_strategy` to `true` -
https://github.com/thoughtbot/factory_bot/releases/tag/v5.0.0.rc2.  If
we want our factories to continue to build valid records we sometimes
need to persist the associations which are created. This change
explicitly sets the `strategy: create` on associations in factories for
which we have failing specs.

Co-authored-by: Adam Mueller <adam@super.gd>
Co-authored-by: Alex Blackie <alex@super.gd>
Co-authored-by: Nick Van Doorn <nick@super.gd>