Check for existence of `product_path` instead of `Spree::Frontend::Engine`
Goal
----

We want to update the Edit Product page to check for existence of
`product_path` instead of `Spree::Frontend::Engine` so that the Edit
Product page would work with a frontend generated by
SolidusStarterFrontend.

Requirements
------------

*Admin can find a link to the product in the Edit Product page if the app
has a frontend*

Given I am an admin of a Solidus app
And the app has a product page i.e. it has a frontend
When I go to the Edit Product page
Then I should see a link to the product page

*Admin won't find a link to the product in the Edit Product page if the
app does not have a frontend*

Given I am an admin of a Solidus app
And the app has does not have a product page i.e. it does not have a
  frontend
When I go to the Edit Product page
Then I should not see a link to the product page