Deprecate Spree::PromotionRule.for
This class method is only called from Spree::Promotion#eligible_rules,
where it adds a `.all` to the method chain. This `.all` causes a request
to the database while making no changes to the array of possibly
preloaded rules. Additionally, it adds a layer of indirection to the
already complex promotion code.

In the specs, this allows us to remove some stubbing.