Fix(OrderMerger): Do not carry line item adjustments to new order
Adjustments are connected to the spree_orders table via two ways:
- via their adjustable's (line item's) order ID
- via their own order ID

If we move a line item with adjustments to new order, the adjustment's
order ID will not be changed to the new order.

Instead of changing the foreign key of an existing line item here, let's
build a new line item with the old line item's attributes.