Add Promotion#activatable?
This checks whether a promotion can at all be activated for the current
order. It does not check whether anything is blocklisted, as that
would prevent a line item to be discountable if any line items on the
order were non-discountable, something I consider a bug with the legacy
system.