Deprecate Promotion#match_policy == any
Using the "any" match policy on promotions creates a lot of difficulty
when working with the promotion system. A recent poll on the Solidus
Slack indicated that most stores do not use these.

It is also from the user's point of view preferable to create separate
promotions for each rule in this case, as this will allow store owners
to set individual start and end dates and so on for each rule.