Add PromotionRules/Actions extension point for preloading relations
Some rules and actions need to preload related records for performance
reasons. Prior to this commit, the promotion handler needed to know
which rules need to preload which relations; with this commit we have a
simple API that helps rule authors define preload relations without
having to modify the promotion handler.