Fix test involving discard + stub_model
For some reason, calling `#reload` on a stubbed model with soft deletion
doesn't work anymore. We didn't dig further but instead changed to a
better practice of testing the actual thing. [`stub_model` is known to
break](https://github.com/rspec/rspec-rails/issues/894) for new
ActiveRecord releases, so it's good to get rid of it progressively.