Fix test on Rails 7
I wasn't able to nail it down, but the combination:

- A `:has_many` association with a scope
- An inverse `:belongs_to` association with `touch: true`
- Using `build_stubbed` Factory Bot strategy for the parent model
- Using Rails' `CollectionProxy#<<`

It raises because the parent model is touched (as stubbed models can't
reach the database). However, using the `build` strategy doesn't seem to
try to touch the record. I time framed the debugging, and I couldn't
isolate it. I was able to track it down to
https://github.com/rails/rails/pull/43358/files#diff-7c2226b7a4aa7f86f7d9e5e16b10686056c02efc725a60e378b5f9d7e52e8403,
though. As it's just a test we can fix by just using `build`, it's not
worth the effort to keep going with the investigation.