Make Omnes the default for pub/sub events
We add an Omnes subscriber class for order mailing work, equivalent to
the legacy subscriber module.

From now on, new applications will use Omnes by default. We also run our
test suite with it, but we still allow running with the legacy adapter
through a `USE_LEGACY_EVENTS` environment variable (and use it in a new
CI job).