Mark failing tests on SQLite
SQLite adapter fails to provide a unique index for the convoluted chain
of ActiveRecord persistence calls that happen on
https://github.com/solidusio/solidus/blob/8b03161a1894a7d10c8d95f62c016da55027d2b0/core/app/models/spree/stock/simple_coordinator.rb#L83.

```
SQLite3::ConstraintException:
UNIQUE constraint failed: spree_shipping_rates.shipment_id, spree_shipping_rates.shipping_method_id
```

That's not a problem on production databases (MySQL & PostgreSQL), but
at least we mark the tests as failing locally. For some reason, we've
only been able to reproduce that on Ruby 3.1.

[skip ci]