Rename spree_products.discontinue_on to spree_products.available_until
In our business, the semantics of availability and discontinuation are
not aligned with Solidus' nomenclature: A product can both be available
and discontinued (i.e. we still have stock of it, but won't be producing
any more).

When communicating the meaning of the discontinue_on column, it's often
hard (because there is a date when we decide to discontinue, and there's
another date when we sell out and the product becomes unavailable).

From reading the code, Solidus' `discontinue_on` is actually a mirror of
`available_until`.

I've added deprecations for usages of `discontinued?` and
`discontinue_on`, and an `alias_attribute` that also fixes queries:

```
➜  sandbox git:(rename-discontinue-on) ✗ rails c
Loading development environment (Rails 6.1.5)
irb(main):001:0> Spree::Product.where(discontinue_on: nil).to_sql
   (1.3ms)  SELECT sqlite_version(*)
=> "SELECT \"spree_products\".* FROM \"spree_products\" WHERE
\"spree_products\".\"deleted_at\" IS NULL AND
\"spree_products\".\"available_until\" IS NULL"
irb(main):002:0>
```

I know this is a somewhat breaking change.