Use non-bang version of `Order#next`
Like exposed with the previous commit, the bang version of `next`
has the drawback that it raises an error when it cannot advance the
order state machine, which may happen especially when the order
state machine has been customized.

In general, restarting the checkout flow for an order should be able
to advance the order to the previous state, but this may not be the
case if some attributes of the order changed before calling this
method and now the order cannot advance anymore. In that situation,
the previous code was going to raise an error but that seemed to be
a bug, since there doesn't seem to be anything wrong with keeping the
order in cart state, if it cannot advance.