Add failing spec for `Spree::Order#restart_checkout_flow`
This new spec fails and shows that with a very simple customization
the method `Spree::Order#restart_checkout_flow` can misbehave when
the order cannot transition forward from the `cart` state by raising
a `StateMachines::InvalidTransition` error.

This happens because of this line from `#restart_checkout_flow`:

    next! if !line_items.empty?

This line assumes that the order can always transition from `cart` to
the next state, if it has line items. The `if` guard exists exactly
for that reason: trying to `next!` an order without line items raises
the above mentioned `StateMachines::InvalidTransition` error.

This line was modified with c9eb45d for no other reason than fixing
some specs, according to the commit message.