Handle empty values as quantity adding to cart via API
Right now, when empty values are passed as quantity in the API call that
creates line items, the line item is populated with quantity 0.

This is not the expected behavior, as it makes no sense to have a line
item with 0 as quantity.

Still, if someone wants to force the value to 0, this is still possible
but 0 needs to be explicitely passed as quantity value. A spec has been
added for this scenario as well.