Fix CSRF vulnerability that allowed (un)finalizing adjustments
As the routes to finalize/unfinalize adjustments were `GET` requests, an
attacker could run them without the user's realization.

We change both routes to be `PUT` requests. We're adding missing tests
for the feature and regression tests for the CSRF vulnerability. We're
explicitly excluding `:show` from the adjustments' routes so that the
meaningful exception `ActionController::RoutingError` is raised when
trying to access via `GET`.

See
[GHSA-8639-qx56-r428](https://github.com/solidusio/solidus/security/advisories/GHSA-8639-qx56-r428)
for details and instructions to reproduce.