Reorganize api shipment creation specs
This commit move the generic specs in the outer context
and keeps the one related to the deprecated variant_id
params in its own context.

It will be easier to remove them when it's time to remove
deprecated code.

This commit also removes a code comment that at the moment
doesn't provide a lot of value.