Make the inventory unit builder configurable
This adds the inventory unit builder (Spree::Stock::InventoryUnitBuilder
by default) to the Spree::Config.stock configuration, so that it can be
overridden.

This is desirable for any store that doesn't have a one-to-one mapping
between purchased variants and the fulfilled inventory units. The
extension solidus_product_assembly heavily overrides this class to
accomplish this, so with this change it will be able to stop decorating
Spree::Stock::InventoryUnitBuilder and instead provide its own.

I also reworked the stock configuration specs a bit. I fixed the
repeated typo, simplified the structure slightly, and ensured they don't
leak constants anymore.