Make the inventory unit builder configurable
This adds the inventory unit builder (Spree::Stock::InventoryUnitBuilder
by default) to the Spree::Config.stock configuration, so that it can be
overridden.

This is desirable for any store that doesn't have a one-to-one mapping
between purchased variants and the fulfilled inventory units. The
extension solidus_product_assembly heavily overrides this class to
accomplish this, so with this change it will be able to stop decorating
Spree::Stock::InventoryUnitBuilder and instead provide its own.

I'll admit that I'm not a huge fan of how I tested the functionality in
the order class, but there's a reason for stubbing the preference like
that. Spree::Config.stock isn't a normal config object, so it doesn't
support our test helpers for stubbing preferences within the context of
only a single test.