Make the availability validator configurable
This makes the Spree::Stock::AvailabilityValidator swappable.

I need to justify using `send` in that test: this method is called by
the order state machine, itself a swappable component. I could imagine
making it a public method, but I don't want to increase the surface area
of Spree::Order's API. Instead, I think it's useful to test this method
in isolation (i.e. separately from the state machine) with the
assumption that it is part of the API that the configured state machine
will call.

I've not added tests for the change to the checkout controller in the
frontend... because it's too hard. The configured class already gets
called a bunch just in setting up these tests, so trying to write a test
that cares that it gets called with the right arguments at that
particular moment did not seem worth the effort given we're deprecated
frontend. Hopefully we can do something better in
solidus_starter_frontend, where we also need to care about whether we're
on a Solidus version that supports this configuration option.