Only install ActiveStorage adapter on supported Rails versions
We only support the ActiveStorage adapter for Rails >= 6.1,
but we install it for any supported Rails versions by default.

If generating a dummy test app for an extension this leads to
failing test runs, because Solidus quits oif the Rails version
is not Rails 6.1 if using the ActiveStorage adapter.

Solving by installing the ActiveStorage adapter only if the Rails
version is >= 6.1.