Update CircleCI configuration
Images prepended with `circleci/` are deprecated [1] in favor of `cimg/`
ones.

Adding the `-browsers` suffix will pull node & selenium, but not the
actual browsers. For that, we need to add the `browser-tools` orb and
install the browser & driver in new steps.

The `-ram` suffix [2] in mysql is no longer supported [3].

[1] - https://discuss.circleci.com/t/legacy-convenience-image-deprecation/41034
for details.

[2] - https://circleci.com/docs/2.0/circleci-images#service-image-variant

[3] - https://circleci.com/developer/images/image/cimg/mysql