Use nested Carmen regions if necessary for state factory
For some countries such as Italy, its states are nested within regions,
and not at the first level of subregions like 'US'.
`country > region > states`
`Carmen::Country.coded('IT').subregions.first.type`

Now nested subregions exist, it will use this instead. This will be
useful for when using the state factory, but also necessary for the
upcoming commit, where if a state is not given, but a country is, for
the address factory, it can automatically use/create the first state of
that country.