Fix stores with no authentication
We deprecated `#try_spree_current_user` in favor of
`#spree_current_user` in #3923 [1]. The former defaulted to `nil` when
no authentication library was present, but the latter was simply not
defined under those circumstances.

This commit restores the old behavior, as the code base works with the
assumption that current user is `nil` when no extension has defined it
(e.g. [2] & [3]).

The issue was raised in the support channel [4].

[1] - https://github.com/solidusio/solidus/pull/3923
[2] - https://github.com/solidusio/solidus/blob/fee8a860c068b6ad050f9b2ff9052aadf9d56174/backend/app/controllers/spree/admin/base_controller.rb#L36
[3] - https://github.com/solidusio/solidus/blob/fee8a860c068b6ad050f9b2ff9052aadf9d56174/backend/app/controllers/spree/admin/cancellations_controller.rb#L33
[4] - https://solidusio.slack.com/archives/C0JBKDF35/p1657030519259159?thread_ts=1657019907.668419&cid=C0JBKDF35