Bump Solidus to 3.0.7
If you look at the last commit [1], you'll see it's bumping to v3.0.6.
However, that commit doesn't represent an actual rubygem version update.
v3.0.6 was released on June 01, 2022 [2].

We did create a v3.0.6 tag [3] on that date. However, the associated
commit [4] was later removed.

We've manually checked that all the commits backported after that date
are still in the branch.

Now, we need to bump twice the code so that we can release v3.0.7.

[1] - https://github.com/solidusio/solidus/commit/63a9ff7f10f8fb9037fedbd358b8f94a3758470b
[2] - https://rubygems.org/gems/solidus/versions/3.0.6
[3] - https://github.com/solidusio/solidus/releases/tag/v3.0.6
[4] - https://github.com/solidusio/solidus/commit/5a500c2fd9885047165ada08a8d339a68cb83fe4