Fix Gemfile order dependency solidus_frontend <> solidus_auth_devise
solidus_frontend tries to load a partial from solidus_auth_devise [1] [2].
However, it provided a fallback [3] when the latter was missing.

Unfortunately, it worked with the assumption that solidus_auth_devise
went below solidus_frontend in the Gemfile. Otherwise, the fallback
template took precedence.

We're now removing the fallback template and rendering the template only
when it's present. This way, it makes no critical which gem goes
first.

[1] - https://github.com/solidusio/solidus_frontend/blob/dae732ed7442a755cc0426cd2afa6ac662de7105/app/views/spree/shared/_nav_bar.html.erb#L4
[2] - https://github.com/solidusio/solidus_auth_devise/blob/bfbf616ef625aa51c6b74040d96998082f736e13/lib/views/frontend/spree/shared/_login_bar_items.html.erb
[3] - https://github.com/solidusio/solidus_frontend/blob/dae732ed7442a755cc0426cd2afa6ac662de7105/app/views/spree/shared/_login_bar_items.html.erb