Deprecate solidus_frontend & allow installing solidus_starter_frontend
We're recommending new stores use solidus_starter_frontend [1] as their
storefront. However, we still want to give the option to install
solidus_frontend for now, as some extensions rely on it. We're modifying
the Solidus installer to allow installing one or the other.

Nonetheless, solidus_frontend is part of the solidus meta-gem [2], which
is what is usually bundled in a new installation. At this point, we
can't remove solidus_frontend from the meta-gem, as users upgrading from
old Solidus versions could see their storefronts suddenly gone. As that
will be a breaking change, we're deferring it until v4.0.

Until that happens, we must prevent users choosing the new frontend from
pulling solidus_frontend. Because of that, we're breaking down the
solidus gem into solidus_core, solidus_api, solidus_backend &
solidus_sample. We leverage Bundler's injector [3] to add dependencies
matching the main solidus gem constraints in the Gemfile. We also use it
to remove the meta-gem. Still, we need to extend the upstream injector
to support the `path:` option, which is required for local testing
(e.g., the sandbox application). We also must break down the meta-gem
when solidus_frontend is installed until we remove the frontend
directory from this repo. That prevents Bundler from resolving it from
there when `path:` or `git:` are used as sources.

As per the logic to detect which frontend gets installed, the first of
the following conditions met wins:

- It's explicitly selected with the `FRONTEND` environment variable
  (useful for sandbox applications).
- It's explicitly selected with the `frontend` option.
- If `solidus_frontend` is in the Gemfile, that's the one used.
- If the `auto_accept` option is given, `solidus_starter_frontend` is
  chosen.
- The user is prompted to select one or the other.

Extensions have been modified to explicitly depend on `solidus_frontend`
on their Gemfile, so the dummy app generator used for testing will work
OOO matching the third condition. As per its sandbox application,
they'll default to `solidus_starter_frontend` because they use
`auto_accept`, but they can select `solidus_frontend` through the
`FRONTEND` variable.

[1] - https://github.com/solidusio/solidus_starter_frontend
[2] - https://github.com/solidusio/solidus/blob/a22723079f88387f53f72b503db875b9d97aa3f5/solidus.gemspec#L27
[3] - https://github.com/rubygems/bundler/blob/master/lib/bundler/injector.rb