Extract Promotion Adjusting from OrderUpdater
The Order Updater does many things, and one thing it does only semi-well
is recalculating promotion adjustments. Promotion adjustments have an
action as a source, but when the action asks its promotion whether it's
eligible for a particular promotable (the promotion adjustment's
adjustable), the promotion will only ask those promotion rules that are
of the `applicable` to the promotable's type. This is not necessarily
true for line item promotion adjustments, for example.

What this commit does is a first step: Extract the current
promotion-related behavior from the order updater. It also extracts the
relevant specs from the order updater.