Fix Gemfile order dependency solidus_backend <> solidus_auth_devise
solidus_backend tries to load a partial from solidus_auth_devise [1]
[2]. However, it provided a fallback [3] when the latter was missing.

Unfortunately, it worked with the assumption that solidus_auth_devise
went below solidus_backend in the Gemfile. Otherwise, the fallback
template took precedence.

We're now renaming the fallback partial and branching to the one in
solidus_auth_devise when present. This way, it makes no critical which
gem goes first.

[1] - https://github.com/solidusio/solidus/blob/a22723079f88387f53f72b503db875b9d97aa3f5/backend/app/views/spree/admin/shared/_navigation.html.erb#L10
[2] - https://github.com/solidusio/solidus_auth_devise/blob/master/lib/views/backend/spree/admin/shared/_navigation_footer.html.erb
[3] - https://github.com/solidusio/solidus/blob/master/backend/app/views/spree/admin/shared/_navigation_footer.html.erb