Move order and item total calculation to OrderAdjustmentsRecalculator
We want to make the contract between the order updater and the promotion
adjuster simpler.

What this does is move the promo total calculation for both items and
the order into the OrderAdjustmentsRecalculator.
This restricts the surface area of
the contract between order updater and promotion adjuster to JUST the
order's and line item's promotion total, which is everything the
taxation code needs.