Reimplement what Adjustment#recalculate in Order Cancellations Adjuster
Spree::Adjustment#recalculate tries to do the same thing for too many
things - it's especially bad with promotions, where it will do wrong
things. I want to deprecate it, so I need to reimplement its
functionality elsewhere.

Tax adjustments are already recalculated in `Spree::Tax::OrderAdjuster`,
so for tax adjustments that code is not called. But for Unit Cancels, we
need to handle this situation.