Fix unit cancellation adjustment amount computation with promos
The test case here had a promotion that actually made the line item more
expensive, which I do not think is a realistic use case. With the spec
setup fixed, we're calculating the last adjustment to be either the
price or the rest of the line item amount.

This will fall flat on its face if we want to short ship actually
discounted items.