Remove PayPal as an option during the installation process
PayPal has stopped supporting its Ruby SDK [1]. From its README:

> This SDK is deprecated. You can continue to use it, but no new
features or support requests will be accepted. For alternatives, please
visit the current SDK homepage on the PayPal Developer Portal

From the PR [2] that introduced the deprecation message:

> There is no team at PayPal that supports this repo right now (and, in
> fact, the documentation for SDK has been removed from the PayPal docs
> site).
>
> My team is in the process of creating replacement SDKs for integrating
> with PayPal, but until those are available our official recommendation
> is to do one of the following:
>
> A) Use the API directly
> B) Continue using the current SDK, but know that no further updates
> will be made outside of necessary security updates

On top of that, the SDK is locked to paypalhttp_ruby lesser than 2 [3],
but Ruby 3 is only supported since that version [4]. That combines to
make it impossible to use Ruby > 3 + Solidus + PayPal extension.

If PayPal restarts giving support to Ruby, we'll add it again as an
option in the installer for solidus_frontend [5].

[1] - https://github.com/paypal/Checkout-Ruby-SDK
[2] - https://github.com/paypal/Checkout-Ruby-SDK/pull/34#issuecomment-1190523120
[3] - https://github.com/paypal/Checkout-Ruby-SDK/blob/7170ecd5656d9e9f96a4d0ce0ee31b59b3a8268b/paypal-checkout-sdk.gemspec#L22
[4] - https://github.com/paypal/paypalhttp_ruby/issues/12#issuecomment-1194594788
[5] - https://github.com/solidusio/solidus_frontend/blob/master/lib/generators/solidus_frontend/install/install_generator.rb