Do not break down meta-gem when solidus_frontend is chosen
In ebfd2ed3105215cdcf54424b3d4a946f0af300ce, we updated the installer to
allow the selection of both solidus_frontend and
solidus_starter_frontend.

When solidus_starter_frontend is chosen, we need to break down the
solidus meta-gem so that solidus_frontend is not pulled.

We also broke down the meta-gem when solidus_frontend was chosen to
avoid source conflicts with the gem in the mono-repo. However, we
removed the `frontend` directory in
c89fbf07a6d063b1a90faa703a17d57101eb75ee. That means that's no longer
needed, as the gem is explicitly added to the Gemfile.