Let bin/rspec automatically detect the subdirectory in which the spec lives
If a spec file is missing the prefix (e.g. backend) `bin/rspec` will
look for an existing file in any of the subdirectories and run rspec
with all matching files.

E.g. calling

  rspec spec/features/admin/orders/order_details_spec.rb:385

is equivalent to calling

  rspec backend/spec/features/admin/orders/order_details_spec.rb:385

This allows copying and pasting RSpec outputs for failing specs and
having them work just by adding `bin/` in front of them.