Improve error messages for a misconfigured test DB
When running bin/build with the default postgres adapter the DummyApp
will interrupt the loading of the spec_helper. This means that the
first spec will report the correct error (ActiveRecord::DatabaseConnectionError)
but each following spec will try to load the spec_helper again and
will report a confusing "Application has been already initialized."
error.

With this change the original ActiveRecord::DatabaseConnectionError
will be reported for every spec, making it easier to fix it.